change to datatypes for numpy > 1.23

+++ src/schnetpack/data/atoms.py
@@ -327,7 +327,7 @@ class AtomsData(Dataset):
 
     def __getitem__(self, idx):
         _, properties = self.get_properties(idx, self.load_only)
-        properties["_idx"] = np.array([idx], dtype=np.int)
+        properties["_idx"] = np.array([idx], dtype=int)
 
         return torchify_dict(properties)
 
@@ -567,7 +567,7 @@ class ConcatAtomsData(ConcatDataset):
 
     def __getitem__(self, idx):
         _, properties = self.get_properties(idx, self.load_only)
-        properties["_idx"] = np.array([idx], dtype=np.int)
+        properties["_idx"] = np.array([idx], dtype=int)
 
         return torchify_dict(properties)
 
@@ -632,7 +632,7 @@ class AtomsDataSubset(Subset):
 
     def __getitem__(self, idx):
         _, properties = self.get_properties(idx, self.load_only)
-        properties["_idx"] = np.array([idx], dtype=np.int)
+        properties["_idx"] = np.array([idx], dtype=int)
 
         return torchify_dict(properties)
 
@@ -670,7 +670,7 @@ def _convert_atoms(
         inputs = output
 
     # Elemental composition
-    inputs[Properties.Z] = atoms.numbers.astype(np.int)
+    inputs[Properties.Z] = atoms.numbers.astype(int)
     positions = atoms.positions.astype(np.float32)
     if centering_function:
         positions -= centering_function(atoms)
@@ -680,7 +680,7 @@ def _convert_atoms(
     nbh_idx, offsets = environment_provider.get_environment(atoms)
 
     # Get neighbors and neighbor mask
-    inputs[Properties.neighbors] = nbh_idx.astype(np.int)
+    inputs[Properties.neighbors] = nbh_idx.astype(int)
 
     # Get cells
     inputs[Properties.cell] = np.array(atoms.cell.array, dtype=np.float32)
@@ -689,11 +689,11 @@ def _convert_atoms(
     # If requested get neighbor lists for triples
     if collect_triples:
         nbh_idx_j, nbh_idx_k, offset_idx_j, offset_idx_k = collect_atom_triples(nbh_idx)
-        inputs[Properties.neighbor_pairs_j] = nbh_idx_j.astype(np.int)
-        inputs[Properties.neighbor_pairs_k] = nbh_idx_k.astype(np.int)
+        inputs[Properties.neighbor_pairs_j] = nbh_idx_j.astype(int)
+        inputs[Properties.neighbor_pairs_k] = nbh_idx_k.astype(int)
 
-        inputs[Properties.neighbor_offsets_j] = offset_idx_j.astype(np.int)
-        inputs[Properties.neighbor_offsets_k] = offset_idx_k.astype(np.int)
+        inputs[Properties.neighbor_offsets_j] = offset_idx_j.astype(int)
+        inputs[Properties.neighbor_offsets_k] = offset_idx_k.astype(int)
 
     return inputs
 
@@ -706,9 +706,9 @@ def torchify_dict(data):
     torch_properties = {}
     for pname, prop in data.items():
 
-        if prop.dtype in [np.int, np.int32, np.int64]:
+        if prop.dtype in [int, np.int32, np.int64]:
             torch_properties[pname] = torch.LongTensor(prop)
-        elif prop.dtype in [np.float, np.float32, np.float64]:
+        elif prop.dtype in [float, np.float32, np.float64]:
             torch_properties[pname] = torch.FloatTensor(prop.copy())
         else:
             raise AtomsDataError(
